DRY dummy data deserialization functions
